Fix bad init value of `cur_thread_cnt`
It was init with zero, which cause first iteration of `pthread_join`
not being executed at first iteration of test, when `cur_thread_cnt`
is zero, `pthread_join` loop inside the test is skipped due to the
condition is never meet (`x` (which is 0) < `cur_thread_cnt` (which
is 0)). And there is a fatal operation without execution of
`pthread_join`, when your `TEST_COUNT` is large, you may creating
new threads before previous test is not done yet. Creating of new
threads is not the fatal error point, the point is that threads
belongs to previous test is possible to modify `idx`, which means
your idx may occur OOB access of `time_res`, which further lead to
segmentation fault.

P.S. This commit explaned why I said mutex may not working in the
commits, it just because of my typo of initial value of for loop.
Whatsmore, `usleep` should kept since TIME_WAIT of TCP stack
still need it (using that time to freeing resources).